requirements.txt
google-genai>=0.3.0
playwright>=1.40.0
